refactor(examples): group examples by service into 6 subfolders (#621)
Move 45 examples under examples/{consensus,schedule,contract,file,network,signing}/ with cleaner action-prefixed names. Each rename is git mv so history is preserved.

Fix 8 broken relative imports in contract/ files (moved one level deeper, paths bumped ./ → ../ and ../src → ../../src):
  contract/ecrecover.js, ecrecover-via-mirror-node.js, create-stateful.js,
  create-simple.js, create-with-value.js, solidity-precompile.js,
  hooks.js, lambda-sstore.js.

No infrastructure changes — tsconfig/Taskfile/run-all-examples already handle one-level subfolders from the account/ commit.

Top-level shrinks from 58 .js files to 13. The remaining are the misc cluster + ContractHelper.js + run-all-examples.js + serialize-deserialize.js.
