+++ examples/consensus/with-admin-key.js
