+++ examples/contract/create-simple.js
@@ -2,9 +2,8 @@ import {
     Wallet,
     LocalProvider,
     ContractCreateTransaction,
-    ContractExecuteTransaction,
     FileCreateTransaction,
-    ContractFunctionParameters,
+    ContractDeleteTransaction,
     ContractCallQuery,
     Hbar,
 } from "@hiero-ledger/sdk";
@@ -14,7 +13,7 @@ import dotenv from "dotenv";
 dotenv.config();
 
 // Import the compiled contract
-import stateful from "./stateful.json" with { type: "json" };
+import helloWorld from "../hello_world.json" with { type: "json" };
 
 async function main() {
     if (
@@ -35,10 +34,10 @@ async function main() {
         provider,
     );
 
-    // The contract bytecode is located on the `object` field
-    const contractByteCode = /** @type {string} */ (stateful.object);
-
     try {
+        // The contract bytecode is located on the `object` field
+        const contractByteCode = helloWorld.object;
+
         // Create a file on Hedera which contains the contact bytecode.
         // Note: The contract bytecode **must** be hex encoded, it should not
         // be the actual data the hex represents
@@ -47,7 +46,8 @@ async function main() {
             .setContents(contractByteCode)
             .freezeWithSigner(wallet);
         transaction = await transaction.signWithSigner(wallet);
-        const fileTransactionResponse =
+
+        let fileTransactionResponse =
             await transaction.executeWithSigner(wallet);
 
         // Fetch the receipt for transaction that created the file
@@ -61,14 +61,6 @@ async function main() {
 
         // Create the contract
         let contractCreateTransaction = await new ContractCreateTransaction()
-            // Set the parameters that should be passed to the contract constructor
-            // In this case we are passing in a string with the value "hello from hedera!"
-            // as the only parameter that is passed to the contract
-            .setConstructorParameters(
-                new ContractFunctionParameters().addString(
-                    "hello from hedera!",
-                ),
-            )
             // Set gas to create the contract
             .setGas(300000)
             // The contract bytecode must be set to the file ID containing the contract bytecode
@@ -79,6 +71,7 @@ async function main() {
             .freezeWithSigner(wallet);
         contractCreateTransaction =
             await contractCreateTransaction.signWithSigner(wallet);
+
         const contractTransactionResponse =
             await contractCreateTransaction.executeWithSigner(wallet);
 
@@ -100,11 +93,10 @@ async function main() {
             // Set which contract
             .setContractId(contractId)
             // Set the function to call on the contract
-            .setFunction("get_message")
+            .setFunction("greet")
             .setQueryPayment(new Hbar(1))
             .executeWithSigner(wallet);
 
-        // Check if an error was returned
         if (
             contractCallResult.errorMessage != null &&
             contractCallResult.errorMessage != ""
@@ -126,60 +118,22 @@ async function main() {
         const message = contractCallResult.getString(0);
         console.log(`contract message: ${message}`);
 
-        // Call a method on a contract exists on Hedera, but is allowed to mutate the contract state
-        let contractExecuteTransaction = await new ContractExecuteTransaction()
-            // Set which contract
+        let contractDeleteTransaction = await new ContractDeleteTransaction()
             .setContractId(contractId)
-            // Set the gas to execute the contract call
-            .setGas(75000)
-            // Set the function to call and the parameters to send
-            // in this case we're calling function "set_message" with a single
-            // string parameter of value "hello from hedera again!"
-            // If instead the "set_message" method were to require "string[], uint32, string"
-            // parameters then you must do:
-            //      new ContractFunctionParameters()
-            //          .addStringArray(["string 1", "string 2"])
-            //          .addUint32(1)
-            //          .addString("string 3")
-            .setFunction(
-                "set_message",
-                new ContractFunctionParameters().addString(
-                    "hello from hedera again!",
-                ),
-            )
+            .setTransferAccountId(wallet.accountId)
             .freezeWithSigner(wallet);
-        await contractExecuteTransaction.signWithSigner(wallet);
-        const contractExecTransactionResponse =
-            await contractExecuteTransaction.executeWithSigner(wallet);
+        contractDeleteTransaction =
+            await contractDeleteTransaction.signWithSigner(wallet);
+        const contractDeleteResult =
+            await contractDeleteTransaction.executeWithSigner(wallet);
 
-        await contractExecTransactionResponse.getReceiptWithSigner(wallet);
+        // Delete the contract
+        // Note: The admin key of the contract needs to sign the transaction
+        // In this case the client operator is the same as the admin key so the
+        // automatic signing takes care of this for you
+        await contractDeleteResult.getReceiptWithSigner(wallet);
 
-        // Call a method on a contract that exists on Hedera
-        const contractUpdateResult = await new ContractCallQuery()
-            // Set which contract
-            .setContractId(contractId)
-            // Set gas to use
-            .setGas(75000)
-            // Set the function to call on the contract
-            .setFunction("get_message")
-            // Set query payment explicitly
-            .setQueryPayment(new Hbar(3))
-            .executeWithSigner(wallet);
-
-        // Check if there were any errors
-        if (
-            contractUpdateResult.errorMessage != null &&
-            contractUpdateResult.errorMessage != ""
-        ) {
-            console.log(
-                `error calling contract: ${contractUpdateResult.errorMessage}`,
-            );
-            return;
-        }
-
-        // Get a string from the result at index 0
-        const message2 = contractUpdateResult.getString(0);
-        console.log(`contract returned message: ${message2}`);
+        console.log("contract successfully deleted");
     } catch (error) {
         console.error(error);
     }
