+++ examples/contract/ecrecover-via-mirror-node.js
@@ -18,7 +18,7 @@ import {
     MirrorNodeContractCallQuery,
 } from "@hiero-ledger/sdk";
 
-import ecrecoverCaller from "./ecrecover_caller.json" with { type: "json" };
+import ecrecoverCaller from "../ecrecover_caller.json" with { type: "json" };
 import { setTimeout } from "node:timers/promises";
 
 dotenv.config();
