+++ examples/file/delete.js
