+++ examples/file/get-contents.js
