+++ examples/network/client-with-mirror-address-book.js
