+++ examples/network/construct-client.js
