+++ examples/network/node-lifecycle.js
