+++ examples/network/registered-node-lifecycle.js
