+++ examples/schedule/network-update.js
