+++ examples/signing/generate-key-with-mnemonic.js
