+++ examples/signing/generate-key.js
