+++ examples/signing/legacy-sign.js
