+++ examples/signing/local-wallet-info-query.js
