+++ examples/signing/local-wallet-sign.js
