+++ examples/signing/multi-node-multi-sig-remove-all.js
