+++ examples/signing/multi-node-multi-sig-remove.js
