+++ examples/signing/multi-node-multi-sig.js
