+++ examples/signing/multi-sig-offline.js
