+++ examples/signing/sign-with-hsm.js
